Added "-daemon" command-line switch to start conkeror in daemon mode
This switch suppresses the default behavior of opening the homepage
("-batch") and calls "conkeror.daemon_mode(1)".  In addition, to
workaround the issue with xulrunner exiting immediately after
processing the arguments, despite daemon mode, if no windows are ever
opened, a temporary frame is created that is closed before it ever
actually displays to the user.