Load the rc file before command-line arguments are processed
If the "-q" option (suppress loading the rc file) is specified, it
must be the first argument.

This change ensures that any URLs to load or JavaScript expressions to
evaluate specified on the command-line are handled uniformly on the
first launch and on subsequent invocations; specifically, they are
handled with any changes in behavior specified by the rc file already
in effect.