Partial rewrite of keyboard binding system
All keys are now bound either by keyCode or by a matching predicate;
binding by charCode is no longer supported.  A keymap now stores the
keyCode bindings in a table such that they can be queried in constant
time (it is still necessary to loop over the predicate bindings,
though).  This should allow key bindings to behave more consistently
across xulrunner versions, as keyCode semantics do not seem to vary as
much as charCode semantics.

A side effect of this change is that all bindings are in terms of key
codes, which are independent of the state of the shift modifier.
Consequently, a key accessible via shift, like ":", must now be bound
as "S-;".

The function kbd now supports receiving a string, rather than just a
single character, as the first argument, which it parses into a key
combination.  Specifically, it supports modifier prefixes, like "C-",
"M-", and "S-", and also supports the symbolic key names of the form
"DOM_VK_*" defined in nsIDOMKeyEvent.  Note that the names should be
used without a "DOM_VK_" prefix, like "C-M-space".  In addition, an
uppercase letter is automatically translated to include the shift
modifier.

Unfortunately, at least on GTK2, Mozilla has a bug regarding the
generation of "keydown" events: if a key is held down and then while
it is still down another key is pressed, no "keydown" event is
generated for the second key press; only a "keypress" event is
generated.  This is due to a sloppy attempt to prevent duplicate
"keydown" events from being generated in case of key repeating.  This
situation is particularly likely to occur when typing multi-key
sequences, which are common in Conkeror.  The result of this bug is
that for some events, no keyCode is provided by Mozilla; only a
charCode is provided.  Fortunately, it is possible to detect the
occurrence of this problem with complete reliability, and then use a
lookup table to convert the charCode to a keyCode.  Note that the
mapping from charCode to keyCode is in general not one-to-one, but
that is not problem.  To avoid encoding information about a specific
keyboard layout in Conkeror, however, an interface is provided whereby
users can merely press arbitrary keys, and Conkeror automatically
builds a mapping table between charCodes and keyCodes for the keys
pressed; this mapping data is saved in the preference
"conkeror.charCodeMappingData", which is read by the keyboard system.
To acces this interface, use M-x charcode-mapping-tool RET.  Note that
this interface runs in a top-level chrome window, to allow it to
capture all key events without interference.