Added support for viewing page source in an external editor
The general-purpose function download_for_external_program was added
to save.js.  This function attempts to fetch a document from the cache
if it is available.  It automatically generates a unique temporary
file name, and calls a callback when the download completes.

The new variable (note: not a preference)
conkeror.view_source_external_editor specifies a string path to the
external editor program to use.  Alternatively,
conkeror.view_source_function specifies a callback function to be
called after the source is contained in a file.  It is called with two
arguments, (file, is_temp), where file is an nsIFile reference to the
file containing the source code, and is_temp is a boolean value
indicating whether the file is a temporary file.  If either of these
variables are set, the internal viewer will not be used.