Fix missing scrollbars due to usechromesheets change
The code in nsDocumentViewer.cpp skips loading the scrollbar
stylesheet if a usechromesheets attribute is specified; since we still
want it to be loaded, we manually specify the scrollbar stylesheet in
the usechromesheets attribute.  Note that the stylesheet specified is
the one used by all platforms except Mac OS X.  This might result in a
changed appearance of scrollbars on Mac OS X.  This problem might be
addressed later.