Initial step of large code reorganization to improve buffer and minibuffer interface
JavaScript code is no longer loaded in frame context.  The file
buffer.js defines the wrapper objects for buffers.  The file
minibuffer.js defines the wrapper objects for the minibuffer.  The
minibuffer can now be used recursively.  Key bindings can now be
buffer-dependent.

The interface to run_hooks is changed: the first argument specifies
the hook variable, and the remaining arguments are forwarded to the
hook functions themselves.  The hook functions are no longer called
with a special `this'; they are instead always called with a null `this'.