Improved directory layout and module system
The @mozilla.org/moz/jssubscript-loader;1 class is now used to load
the modules, rather than get_chrome_contents combined with eval.  This
allows the correct line numbers and file names to be associated with
functions for debugging purposes.

The modules are now located in modules/, rather than
conkeror/content/, such that they are better distinguished.  The
subdirectories of the top-level conkeror directory, namely branding,
locale, and content, were moved to be top-level directories themslves.
The contents of modules/ can be accessed as
chrome://conkeror-modules/content/.

The function `load_module' loads the specified file from the modules
directory.  Additionally, it adds the name of the module to the
modules_loaded variable.  The function `require' can be used to ensure
that a module is loaded, and provides a higher-level interface to
`load_module'; if the specified module name is not found in
modules_loaded, load_module is invoked to load it.  This will allow
individual modules to state their dependencies directly.