build.sh uses CONKEROR_APP_NAME to set application name
If the environment variable CONKEROR_APP_NAME is set, build.sh changes
the application name (Name=) line in application.ini in the built
xulrunner app.  The application name determines the final component of
the path in which the application is installed, it determines where
the profile directory is located.  Furthermore, the application name
determines the set of existing browser instances that are checked to
see if a remote command can be issued instead of starting a new
process.  Using a different name for a second copy of Conkeror allows
that copy to be tested without affecting other copies.