Add new popup infrastructure, change frameset.js to use it
This minimal popup infrastructure provides a simple interface for
adding a new popup without having to manually perform the DOM
manipulations.  Additionally, it automatically closes an existing
popup before opening a new one, to avoid running in the GTK keyboard
grab bug and also to ensure they do not overlap.