Revamp of URL opening/following system
Use the default_browse_targets (defined in modules/browser_buffer.js)
array to specify the behavior of the prefix argument for open and
follow commands; the first element of the array is used by default,
and each C-u moves to the next element.

Use browser_default_open_target (defined in modules/browser_buffer.js)
to specify how to handle new window open requests.  It is by default
set to OPEN_NEW_BUFFER, which results in new window requests being
diverted to a new buffer that is then selected.