Shell command execution (by itself, on URLs, and on temporary downloaded copies of files)
The shell-command command (bound to M-S-1) just allows a normal shell command to be run.

The command browser-element-shell-command-on-url (bound to X) allows a shell
command to be run on the URL of a browser element.

The command browser-element-shell-command (bound to x) allows a shell
command to be run on the temporary downloaded copy of a browser
element.

All of these shell commands are run from the current directory of the
buffer, which is a concept added in this changeset as well.  The
current directory of a buffer is by default the Conkeror default
directory, which is generally the user's home directory.  The current
directory can be changed using the change-current-directory command.
The -cwd <dir> command-line argument sets the current directory for
any new buffers created by subsequent command-line arguments.  New
buffers inherit the current directory of the buffer that "caused"
their creation, which will generally be the previous selected buffer.