modules/content-buffer.js (go_up): behave as expected
Now the go-up command behaves in the expected way, implemented
in a robust way.  If the current url has a query or params, they
will get trimmed.  If it has a filename part, up is the current
directory.  If it does not have a filename part, up is the parent
directory.