Cleanup definition of builtin cmd_ commands, support negative prefix argument
The array of strings `builtin_commands' in utils.js defines the names
of builtin commands that do not accept a count argument.  The array
`builtin_commands_with_count' (also in utils.js) defines the names of
builtin commands that accept a prefix argument.  If an element of the
array is a string, it defines a command that accepts only a positive
count; otherwise, the element must itself be a two-element array of
strings, specifying a pair of opposite command names.