Add basic support for SSL certificate management
The new file module ssl.js defines the command ssl-add-exception that
opens the built-in dialog for adding an exception for the SSL
certificate for a particular site.  Error pages rather than alert
dialogs are now enabled by default, and the error page for an invalid
certificate now references this command for adding an exception.

Additionally, the command ssl-certificate-manager was added.  This
command opens the built-in SSL certificate manager, which can be used
to import new certificate authority certificates and remove exceptions
added by ssl-add-exception, among other things.