Change to coroutine infrastructure
The module coroutine.js defines a coroutine library for JavaScript
based on the JavaScript 1.7 generators.

The interactive system has been very significantly simplified;
interactive methods have essentially been removed.  All interactive
commands are now defined in terms of a single function, which may
optionally be a coroutine, that is called with the interactive context
as the only argument.