removing sticky-modifiers code.
This way of doing sticky modifiers is flawed in that to make it work
with at any point in a key sequence, special handling would be
necessary in key_press_handler.  Yet fundamentally, sticky-modifiers
is an accessibility feature that should be handled at the level of
the OS, rather than in the application.