Implemented new, more powerful framework for spawning proceses
This new framework allows starting processes in a specific working
directory, allows arbitrary file descriptor redirections, and does not
depend on multi-threading for notification of when the process exits.
Instead, it uses an external helper program which it talks to over a
TCP socket.  In order to run external programs, it is necessary to
build the external helper program by running: make.  Currently the new
framework does not work with MS Windows.