minibuffer-completion.js: prefix_completer now uses the common prefix as default completion
This fixes an annoying behavior with using prefix completion with
match_required enabled.  Previously, if you typed e.g.  M-x
shell-commm TAB, the input area would complete to "shell-command", but
"shell-command" would not be selected as the default completion; as a
result, if you then hit RET, you get the message "No match", because
there is no selected completion.  This problem is now fixed.