minibuffer: better support for input methods
The minibuffer no longer relies on keypress event handlers to add
characters to the minibuffer.  Instead, text is normally added to the
minibuffer input field using the default handlers, which is what is
done unconditionally by input methods anyway.  That way, a single
handler for the "input" event can be used to detect changes to the
minibuffer text by the default Mozilla key handler.