contrib/run-conkeror: a launch script that can be symlinked to
run-conkeror is a launch script for systems that have a basic gnu
installation (bash, grep, &c) and support symlinking.  To use it,
create a symlink to it in some directory in your PATH.  It will chase
the symlink to find itself and your conkeror installation, as well as
look in your $HOME/.gre.d and /etc/gre.d to locate your xulrunner
installation.  Armed with these two pieces of information it will
launch conkeror.

For improved startup speed, the user can easily write his/her own
script with paths hardcoded in, but this symlink-chasing script should
suffice for most.