isearch-continue-forward/backward now also work non-interactively
isearch-continue-forward and isearch-continue-backward can now be called
when there is not an open isearch minibuffer (but an isearch has been
completed in this buffer).  This is modelled after the `n' and `N' commands
in vi(m).

new user variable: isearch_keep_selection.  default is false.  when true,
the selection will always be left visible upon completion of an isearch.