key_press_handler: reset active keymap for null bindings
this is a bugfix.  the bug affected hook bindings with no command or
keymap, such as the browser-object-class bindings in the `*' keymap.
After hitting `* M' for example, for MathML nodes, this bug resulted
in conkeror looking for a sub-binding of `* M' rather than reverting
to top_keymap for the next keystroke.