context_case functionality for key lookups
A key definition can use a context_case to dispatch on the interactive
context to provide a command during key lookup.  This can be used to have
vi-like "shortcut" sequences.  The first time you press a context_case key
the browser-object prefix command is run, and the second time, an action
is run.  This allows [ and ] to be browser-object prefix commands, but
[[ and ]] to perform follow-previous and follow-next.