Optionally display a URL panel when selecting from hints.
When hints_display_url_panel is true and an item is being selected
using the hinting system, a panel is displayed above the minibuffer
showing the URL of the item that is being selected.  This is useful for
confirming that the correct link is selected and that the URL is not
undesirable in some way.