reverting minibuffer-read stores match in history when match_required
Explanation from jbms:

This patch does not work because there is no guarantee that the match value
that is returned to the caller is actually a suitable value for the user
to have typed, since getting a value from the history is equivalent in
terms of the completion system to the user having typed it.  In
particular, the match value might not even be a string.  Notice that
with this patch, if you use C-x b, pick something, and then open C-x b
again, and look at the history, you'll see that [Object object] is
listed, because with switch-to-buffer, the match value is the actual
buffer itself, not a string.

I agree that what you want is a useful feature in some cases like when
completing a URL, but I think it would require some change to the
completion system in order for it to work.  Even then, it is not clear
how to make it work with switch-to-buffer.