english prompts for read_browser_object
interactive takes a $prompt keyword to give an english prompt to use
with commands.

I.command is now a reference to the whole command definition, instead
of just the command name.

read_browser_object now takes an interactive context object as an arg,
rather than being a member of interactive_context and using the `this'
variable.

new browser object class, `browser_object_file' for reading a filename