xkcd-mode: no longer insert title by default
The variable xkcd_add_title now controls this behavior.  To make xkcd-mode
add the title text below the comic, put this in your rc:

   xkcd_add_title = true;

This change was made because it is a reasonable expectation that under
default configuration, conkeror should display all web pages as
intended by their authors.  Since xkcd-mode has an additional feature
(follow-next and follow-previous patterns for navigating xkcd) which
is non-breaking, it seemed best to keep this page mode as one that
loads by default, but to make its content-editing feature one that
must be explicitly enabled by the user.