4chan.js: Adding quoted message preview on mouse over.
When hovering ">>1388818" links with the mouse, a floating preview of the
message that has been quoted shows up, much like the image previewing that was
already available. This allows the user to more easily follow the thread
activity, not having to click and scroll back down again.

I'll still have to figure out how to make any potentially attached image of the
quoted post show up in the preview, as the hideous HTML coding on 4chan makes
this a big issue.