utils.js: Adding function ajax_request.
Given a URI and a callback function, ajax_request makes an asynchronous HTTP
request to the URI.  The default request method is GET, but the user may freely
choose to use POST and post some data as well.  It doesn't handle server errors
or anything - this is left for the user to implement.