4chan.js: Fixing bug in chan-(next|previous)-thread.
Previously, we used getElementsByTagName('HR') to determine the next or previous
thread. This didn't work out so well, since there are HR elements on the
imageboard which don't denote delimiters of threads.

This change uses an XPath expression which is separated into one expression
matching the very first thread (right after the form) and one matching every
other thread.