4chan.js: Removing C-c C-c to chan-post-reply.
The reason for removing this keybinding is that people may be hitting C-c C-n or
C-c C-p rapidly, sometimes accidentally hitting C-c C-c. Leaving the interactive
command to let users bind it as they wish. Of course, C-c C-c still the default
binding for submitting forms, so as long as some field in the reply form is
focused, it will work.