Script I used to make new releases from git.
The script is probably not that useful anymore in its current state
since it was targeted on the debian directory not being part of the
upstream repository. But since it can be a base for a new release script
and it was in my old Mercurial repository, I commit it here, too.