new-tabs.js: Added command "toggle-tab-bar".
This command toggles the visibility of the tab bar. We should probably think
about some way to let the user decide whether she wants the tab bar to be
initially shown when Conkeror is started or not. I tried to implement this as a
variable, but didn't quite manage to get it right.