define_webjump: replace $no_argument keyword with $argument
The values of the $argument keyword have the following meanings:

  When true, the webjump requires an argument.

  When false, the webjump does not require an argument,
  and minibuffer completion will not add a space.

  When "optional", the webjump does not require an argument,
  but minibuffer completion will still add a space.

Function webjumps have a default $argument value of true, meaning
they require an argument unless specified otherwise in the call
to define_webjump.

String webjumps work as follows:

  If the string contains the substitution pattern %s, $argument
  is set to "optional", and an alternative no-arg webjump is auto-
  generated by trimming the path from the url.

  If the string does not contain %s, $argument is set to false.

Array webjumps always have $argument = "optional".  The first string
in the array is the main webjump, and the second is the no-arg
alternative.