beginnings of tests for conkeror's commandline UI
tests/shell-nix/commandline.sh depends on finding shunit2 in
the PATH.  You will need to download that program, and mark it
executable, and place it in a directory in your PATH to run
these tests.  shunit2 is not executable out-of-the-box, because
you source it in order to run it.  However, this requirement
seemed preferable to bundling shunit2 with conkeror, or hard-coding
an arbitrary path to shunit2.