get_current_profile: rewrite & fix
The selectedProfile attribute is not in fact the "current profile".
In fact, the name of the current profile is not made available by
xulrunner, so the method needed to get the name of the current
profile is to compare the current profile-directory against the
directories of all known profiles.  Anonymous profiles are possible
in mozilla, so the information provided by this function should
not be depended upon for anything beyond things like window titles
and mode-line widgets.