mode-line: provide a widget to show count of buffers currently loading
This provides a count like (2 loading) in the mode-line.  When no
buffers are being loaded, the widget is blank.  This is useful as a
notifier when a buffer is visiting a web page that loads slowly.

I enable this and the buffer count with:
add_hook("mode_line_hook", mode_line_adder(loading_count_widget), true);
add_hook("mode_line_hook", mode_line_adder(buffer_count_widget), true);