Adding a ~ to the Debian release number of the nightly builds
Since the nightly builds are built early in the day, they should have a
lower version respectively release number than any potential normal
package releases on this day to ensure that those are automatically used.

This done by preceeding the Debian release number of the nightly builds
with a ~ which causes the release numbers as being always lower than
normal release numbers.