Changed the build id to contain username and hostname instead of commit id.
Changed the build id generated during Debian package building to contain
username and hostname instead of commit id. Together with the unix
timestamp, this still should be unique enough.

This is necessary since we want to be able to build Debian packages from
Debian source packages which don't contain the .git directory or any
other information stored in the repository.