webjump: use $alternative keyword; array form no longer accepted
The alternative url (which is used when no argument is supplied by the
user when invoking a webjump) should now be specified explicitly using
the $alternative keyword to define_webjump().  The previous method of
specifying the handler as an array of two strings was peculiar and
will no longer be supported.

Prepared for allowing search engine webjumps to specify an alternative
url by allowing function handlers to have an alternative url.

Expanded the comment describing the parameters to define_webjump().

The supplied webjumps that have alternative urls are changed to use
the new style.