search-engine: provide alternative urls for search engine webjumps
The alternative url is visited when the user provides no arguments to
the webjump; thus the webjump can function both as a way to invoke a
search and as a bookmark.

The alternative may be specified to define_search_engine_webjump().
If no alternative is supplied, conkeror will try to guess an
appropriate url.  The guess uses the text/html url for the search
engine and trims off the path and any "search." prefix on the domain.
This works for all the supplied search engines.

It may be better to attach the alternative to the webjump object and
to have the alternative logic in getWebJump().  Then it need not be in
the handlers specified in define_webjump() and
define_search_engine_webjump().