download manager: show display name, allow target file name to change
The download manager will now update the display of the target file
name each time there is a change in download state.  If the display
name differs from the (basename of) the target name, it is shown
parenthetically.  Changes to the names in hooks are thus reflected in
the display.

The download manager handling of the target file is changed.  It is
now maintained explicitly and can be specified in register_download.
This is necessary because the targetFile property is immutable.

These changes allow, for instance, downloads to a file with a
temporary name that is changed to the final name when the download is
finished.