save_uri: optionally use a temporary download file
This is useful when it is undesirable for some other tool to attempt
to open the downloaded file before it is complete.

If $temp_file is specified then the uri is downloaded to that file and
then renamed to the output_file when the download is finished.  If
$temp_file is the value true then a new temporary file in the same
directory as output_file is used.

The download manager indicates this by showing the temp_file name with
the output_file name following parenthetically.  When the download is
complete and the file is renamed, the download manager will update to
show the new name.

Note that this save_uri usage of the term "temporary" is different to
the download-manager usage; in that case the file is deleted, not
renamed.