New module to define webjumps for git repository summaries.
A webjump to access repositories at a gitweb server can be defined
using define_gitweb_summary_webjump.  The alternative url for the
webjump can be a particular repository, the repository list or user
defined.

The default completer for the webjump will complete against the
repositories available at the gitweb server.  To enable this, the user
must specify where to store the OPML data from the server (using
either $opml_file or gitweb_webjumps_opml_directory) and fetch the
data once using gitweb-webjump-get-opml.

Examples:

In your rc have:

  require("gitweb-webjump.js");
  gitweb_webjumps_opml_directory = get_home_directory();
  gitweb_webjumps_opml_directory.appendRelativePath(".conkerorrc/gitweb-webjumps-opml");

  define_gitweb_summary_webjump("gitweb-ko", "http://git.kernel.org");
  define_gitweb_summary_webjump("gitweb-cz", "http://repo.or.cz/w");

You can now use the following webjumps:

  gitweb-cz conkeror
  gitweb-ko git/git

If you create the directory ~/.conkerorrc/gitweb-webjumps-opml/ and in
conkeror use M-x gitweb-webjump-get-opml and select gitweb-cz then,
once the download is finished, completions will be available for that
webjump.

With thanks to John Foerch for suggestions.