Mentioning new and old name of conkeror-spawn-helper in package description
Luca Capello noted that a search for "conkeror-spawn-helper" in the
Debian package repository yields no results. Now, the new
conkeror-spawn-helper as well as the old spawn-process-helper name are
mentioned in the package description. (Searching for the old name
already found the package since it's part of the package name in one
string.)