Optionally don't accept partial matches as webjumps.
A variable webjump_partial_match (default true) selects whether a
unique prefix is accepted as a webjump.  The user may want to set this
false because webjumps can get in the way when using the default
search feature.  Webjumps are easy to complete, so partial matches are
not that important.