keydown and keyup events are blocked from content
This fixes problems with sites like slashdot and yahoo which have
event listeners on keydown and/or keyup events, causing keypresses in
conkeror to result in multiple things happening at once (a conkeror
command and an event handler in the content page), leading to conflict
and hard-to-predict behavior.

This patch means that keydown and keyup event handlers in content will
no longer work at all.  Conkeror's input-mode mechanism is not
currently capable of supporting simple modification of
quote-input-mode and quote-next-input-mode to allow for these events
to be passed through to content, but it seemed more expedient to fix
the bug than to delay fixing it until the input-mode system can be
rewritten to support the corner case.