buffer.opener: new property
The property buffer.opener exists from the creation of a buffer
until create_buffer_hook has been run, then it is deleted.  It
is a reference to the buffer or other object that caused the
current buffer to be created.  This property can be used for
inheriting properties from the "opener" such as the working
directory (cwd).  For example:


/**
 * This bit of code is how auto-cwd and cwd-inheritance can be done.
 * Auto-cwd means that buffers created from a command-line action will
 * get their cwd set to the working directory of the command-line
 * invocation.  Cwd-inheritance means that new buffers will inherit
 * their cwd from the buffer that caused them to be created.  For
 * auto-cwd, buffer.opener will be `instanceof interactive_context'.
 * For cwd-inheritance, buffer.opener will be `instanceof buffer'.
 */
function cwd_setup (buffer) {
    if (buffer.opener && buffer.opener.local) {
        if (buffer.opener.local.hasOwnProperty('cwd')) {
            buffer.local.cwd = buffer.opener.local.cwd;
        }
    }
}
add_hook('create_buffer_hook', cwd_setup);