remove contrib/hilighting
This approach to highlighting ran into a dead end.  The idea with adding markup
in an external program was that the program would send html back to conkeror,
and conkeror would substitute the html in place of the original text.  However,
because the buffer will have the content-type text/plain, the html will not be
interpretted.  It will just inlined literally.  The content-type of the buffer
cannot be changed.

There are two obvious ways to hack around this problem and still have an external
highlighter:

 1) Have conkeror read back the output of the highlighter, parsing out the markup
    and running the appropriate DOM methods to color the text.

 2) Make a new buffer with content type text/html.

Both of these are clear kludges.  Option 1 effectively doubles the amount of work
and introduces a second layer of parsing, presumably of some intermediate language
that would be invented just for this purpose.  (yuck)

Option 2 would be redundant of, and less convenient than our existing ability to
use the view-source command to send the page into emacs.  I.e, it would not be
seamlessly integrated into the original buffer: instead of having an extra emacs
window to close, you have an extra buffer to close.  Furthermore, it would be less
convenient than emacs for no benefit because we would still have to create
highlighting patterns for every existing language, which emacs can already do.

Therefore, we are going to remove this code from the conkeror repository, and plan
that anyone who wants syntax highlighting built into conkeror will write it in
javascript without introducing dependencies on external programs.  It should be
noted that, given the existence of the view-source command and its ability to send
a document to emacs, the need for a built-in syntax highlighter is not a pressing
one, and would inevitably be introducing an extra feature into conkeror that other
programs can do better.