Move MPL from file debian/copyright to debian/MPL
This allows us to compress the ca. 30kB MPL down to less than 10kB in
the binary packages. The xulrunner-1.9 package does the same, except
that the MPL is already as single file in the .orig.tar.gz.

Compressing COPYRIGHT instead is no option because it also contains
the GPL and LGPL which is available on all Debian systems anyway and
therefore should not be included in the binary packages.

For more details and the background of these considerations, see
http://www.debian.org/doc/debian-policy/ch-docs.html#s-copyrightfile

Additionally, the package version has been updated for the next
package.