index-webjump: New module to define webjumps for index pages.
Index webjumps provide convenient access, with completion, to a set of
web pages that are indexed (referenced) from another page.  For
instance, webjumps for documentation (e.g., git, ImageMagick) are easy
to define.

A webjump to access URLs referenced from an index page is defined
using define_xpath_webjump.  An xpath expression is used to extract
the indexed URLs and the anchor text; this provides completion for the
webjump.  The completion must be enabled using webjump-get-index once
for each index webjump.

This module also subsumes define_gitweb_summary_webjump, which results
in changes to how gitweb webjumps are set up.