rewrote `unfocus' command
unfocus_hook: new hook, called by unfocus. see below.

the behavior of unfocus is based on the situation.  if a hyperlink
or form field is focused, it unfocuses that.  otherwise, if there
is a selection, it deselects.  otherwise, it unfocuses any frame or
plugin, returning focus to the top frame.  in this third case, it
calls the buffer-local hook unfocus_hook, which gives page-modes
the opportunity to focus something other than the top frame. (This
feature is used by gmail-mode.)