fix issue180, misreported key sequence with key aliases
When a key alias was defined, such as C-m aliased as return, although
Conkeror handled the key correctly internally, a wrong key sequence would
be reported in the minibuffer if you hit, for example, C-x C-m.  It would
be reported as `C-x C-m return'.  This patch fixes that.

Also minor style/whitespace cleanup surrounding the bugfix area.